Refactor dependency management by implementing abstract classes for PyPi client, dependency tree builder, and repo downloader; update main module to utilize/give an example of usage for the new abstractions.
